Rename files and update CategorySection which used to be itemCard
