+++ src/App.js
@@ -1,6 +1,6 @@
 import logo from './logo.svg';
 import './App.css';
-import Home from './components/home';
+import Home from './components/Home';
 
 function App() {
 	return (
